Fix messaging timing
Ignore the ACK to a SDP offer, which may come before or after the event
containing the SDP answer, which has the nice side effect that the
sending of the SDP offer now returns a Promise with the SDP answer. This
allows simplification of minni-janus plugin implementations.
It is also a breaking change to minnie-janus plugin implementations.

See https://github.com/michaelfranzl/minnie-janus/issues/2